fix: #100 make new EndStopView
